% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adjacencyGraph}
\alias{adjacencyGraph}
\title{Create an adjacency graph from a document term matrix}
\usage{
adjacencyGraph(dtm, measure = "cosine")
}
\arguments{
\item{dtm}{a document term matrix, either or not in the dtm format from the `tm` package}

\item{measure}{the measure to calcualte adjacency. Currently supports cosine and conditional probability}
}
\value{
A graph in the Igraph format in which edges represent the adjacency of terms
}
\description{
Create an adjacency graph from a document term matrix
}

