% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.freq.R
\docType{data}
\name{animals.freq}
\alias{animals.freq}
\title{Frequency of Animal Responses}
\format{
animals.freq (vector, length = 367)
}
\usage{
data(animals.freq)
}
\description{
Frequency of animal responses from Christensen & Kenett (2019). These frequencies
are used to generate data in the \code{\link[SemNeT]{sim.fluency}} function.
}
\examples{
data("animals.freq")
}
\keyword{datasets}
