% Generated by roxygen2 (4.0.1): do not edit by hand
\name{arrangeIsolates}
\alias{arrangeIsolates}
\title{Organize isolates next to network}
\usage{
arrangeIsolates(g)
}
\arguments{
\item{g}{A graph in the Igraph format, which has the g$layout attribute (set by using the layout. functions)}

\item{fontsize_multiplier}{The wordlayout function takes fontsize into account to determine overlap. As a temporary solution, increasing font size decreased label overlap (but messes up the network if done extremely)}
}
\value{
a network layout, to be assigned to g$layout
}
\description{
Under development (if it proves usefull). Sometimes ignoring isolated words might seem bad, but having them placed randomly within and around the network is just plain annoying. This is an attempt to combine both sentiments.
}

