% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoConverge.R
\name{autoConverge}
\alias{autoConverge}
\title{Autmated Converge Responses}
\usage{
autoConverge(rmat)
}
\arguments{
\item{rmat}{A semnetcleaner filtered response matrix}
}
\value{
The response matrix with the \strong{word} column merged and the \strong{replace}
column removed for all variables
}
\description{
Automated \link[SemNetCleaner]{converge} function merging of columns of binarized response data with another
}
\examples{
\donttest{
convmat <- autoConverge(rmat)
}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
