% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin2resp.R
\name{bin2resp}
\alias{bin2resp}
\title{Binary Responses to Character Responses}
\usage{
bin2resp(rmat)
}
\arguments{
\item{rmat}{A binarized response matrix of verbal fluency or linguistic data}
}
\value{
A list containing objects for each participant and their responses
}
\description{
Converts the binary response matrix into characters for each participant
}
\examples{
charmat <- bin2resp(convmat)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
