% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{chi2}
\alias{chi2}
\title{Compute the chi^2 statistic for a 2x2 crosstab containing the values
[[a, b], [c, d]]}
\usage{
chi2(a, b, c, d, yates_correction = rep(F, length(a)), autocorrect = F)
}
\description{
Compute the chi^2 statistic for a 2x2 crosstab containing the values
[[a, b], [c, d]]
}

