% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{chi2_wordassociations}
\alias{chi2_wordassociations}
\title{Get wordassociations based on ratio and chi-2}
\usage{
chi2_wordassociations(dtm, odds.ratio.thres = 1, chi.p.thres = 0.05,
  fisher.p.thres = 0.05, smoothing_parameter = 1, use.wordcount = T,
  return.graph = T)
}
\arguments{
\item{dtm}{the main document-term matrix}
}
\value{
A dataframe representing and edge.list
}
\description{
Calculate wordassociations (x -> y) as the ratio of the p
}

