% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{chi2_wordassociations_alternative}
\alias{chi2_wordassociations_alternative}
\title{Get wordassociations based on ratio and chi-2}
\usage{
chi2_wordassociations_alternative(dtm, ratio.thres = 1, p.thres = 0.05,
  smoothing_parameter = 1, return.graph = T)
}
\arguments{
\item{dtm}{the main document-term matrix}
}
\value{
A dataframe representing and edge.list
}
\description{
Calculate wordassociations (x -> y) as the ratio of the p
}

