% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nets.R
\name{compare_nets}
\alias{compare_nets}
\title{Plots Networks for Comparison}
\usage{
compare_nets(..., input_list = NULL, title = NULL, plot.args = list())
}
\arguments{
\item{...}{Matrices or data frames of network adjacency matrices}

\item{input_list}{List.
Bypasses \code{...} argument in favor of using a list
as an input}

\item{title}{Character vector.
Characters denoting titles of plots}

\item{plot.args}{List.
Arguments to be passed along to \code{\link[SemNeT]{plot.SemNeT}}}
}
\value{
Plots networks
}
\description{
Uses \code{\link[SemNeT]{plot.SemNeT}} to plot networks.
Accepts any number of networks and will organize the plots
in the number of side-by-side columns using the heuristic of taking the square root of the number of 
input and rounding down to the nearest integer (i.e., \code{floor(sqrt(length(input)))}).

\strong{Examples}
\itemize{
\item{3 networks:}
{1 x 3}
\item{6 networks:}
{2 x 3}
\item{9 networks:}
{3 x 3}
}
}
\examples{
# Simulate Datasets
one <- sim.fluency(10)
two <- sim.fluency(10)

# Compute similarity matrix
cos1 <- similarity(one, method = "cosine")
cos2 <- similarity(two, method = "cosine")

# Compute networks
net1 <- TMFG(cos1)
net2 <- TMFG(cos2)

# Compare networks
compare_nets(
   net1, net2, title = c("One", "Two")
)

# Change node colors
compare_nets(
   net1, net2, title = c("One", "Two"),
   plot.args = list(node_color = "blue") 
)

# Change edge colors
compare_nets(
   net1, net2, title = c("One", "Two"),
   plot.args = list(edge_color = "blue") 
)

}
\references{
Epskamp, S., Cramer, A. O. J., Waldorp, L. J., Schmittmann, V. D., & Borsboom, D. (2012).
qgraph: Network visualizations of relationships in psychometric data.
\emph{Journal of Statistical Software}, \emph{48}, 1-18.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
