% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semantic Network Cleaner.R
\name{converge}
\alias{converge}
\title{Convergence Function}
\usage{
converge(word, replace)
}
\arguments{
\item{word}{The column name that will incoporate the \strong{replace} column's binarized responses (must be characters)}

\item{replace}{The column name that should be merged with the \strong{word} column (must be characters)}
}
\value{
The response matrix with the \strong{word} column merged and the \strong{replace} column removed
}
\description{
Merge a column of binarized response data with another
}
\examples{
\dontrun{

rmat <- converge("cat","abyssinian")
}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
