% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convmat.R
\docType{data}
\name{convmat}
\alias{convmat}
\title{Converged trial data of verbal fluency responses}
\format{A 15x95 response matrix}
\usage{
data(convmat)
}
\description{
Converged trial data of animal verbal fluency responses. The columns are responses and the
rows are the participants.
}
\examples{

data(convmat)
}
\keyword{datasets}
