% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{coOccurenceNetwork}
\alias{coOccurenceNetwork}
\title{Create a co-occurence network}
\usage{
coOccurenceNetwork(x, measure = "cooccurence",
  doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"),
  word.col = getOption("word.col", "word"))
}
\arguments{
\item{x}{either a document-term matrix (DTM) or a tokenlist.}

\item{measure}{the measure to calcualte adjacency. Currently supports cosine and conditional probability}
}
\value{
A graph in the Igraph format in which edges represent the adjacency of terms
}
\description{
Create a co-occurence network
}

