% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{cosine}
\alias{cosine}
\title{Cosine Similarity}
\usage{
cosine(data, addConstant = 0)
}
\arguments{
\item{data}{A binarized dataset of verbal fluency or linguistic data}

\item{addConstant}{A constant to be added to the cosine similarity matrix}
}
\value{
A cosine similarity matrix
}
\description{
Computes cosine similarity
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
