% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semantic Network Cleaner.R
\name{equate}
\alias{equate}
\title{Equate Group Responses}
\usage{
equate(rmatA, rmatB)
}
\arguments{
\item{rmatA}{Response matrix for group 1}

\item{rmatB}{Response matrix for group 2}
}
\value{
A list of responses matched for group 1 (rmatA) and group 2 (rmatB)
}
\description{
An automated cleaning function for matching groups' responses
}
\examples{
\dontrun{

groupA<-read.csv(file.choose(),header=FALSE,sep=",",as.is=TRUE)
groupB<-read.csv(file.choose(),header=FALSE,sep=",",as.is=TRUE)

rmatA<-semnetcleaner(groupA)
rmatB<-semnetcleaner(groupB)

groups_resp_match<-equate(rmatA,rmatB)


}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
