% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_groups.R
\name{extract_groups}
\alias{extract_groups}
\title{Extract Groups from \code{\link[SemNetCleaner]{textcleaner}}}
\usage{
extract_groups(object, groups, type = c("fluency", "free"))
}
\arguments{
\item{object}{Matrix, data frame, or object from \code{\link[SemNetCleaner]{textcleaner}}}

\item{groups}{Vector.
Specify groups using character or numeric values}

\item{type}{Character.
Whether data are from a \code{"fluency"} task or
\code{"free"} association task}
}
\value{
Returns a list containing the cleaned data split by groups
}
\description{
Extract groups from a \code{\link[SemNetCleaner]{textcleaner}} object
}
\examples{
# Openness to experience data


}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
