% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_flow.R
\name{forward_flow}
\alias{forward_flow}
\title{Forward Flow}
\usage{
forward_flow(
  response_matrix,
  semantic_space = c("baroni", "cbow", "cbow_ukwac", "en100", "glove", "tasa", "all"),
  min_response = 10,
  task = c("free", "fluency"),
  prompt_word = NULL,
  cores
)
}
\arguments{
\item{response_matrix}{A data frame containing the set of responses to calculate forward flow on.  Must have id column, followed by all responses for that observation in wide format.}

\item{semantic_space}{The semantic space used to compute the distances between words.}

\item{min_response}{A numeric indicating the minimum number of valid responses needed to calculate forward flow on.}

\item{task}{Type of semantic task}

\item{prompt_word}{If prompt_included = FALSE, the prompt word given to participants.  Will be used as the first word in the forward flow string.}

\item{cores}{Numeric.
Number of computer processing cores to use for bootstrapping samples.
Defaults to \emph{n} / 2 total number of cores.
Set to any number between 1 and maximum amount of cores on your computer
(see \code{parellel::detectCores()})}
}
\value{
A tibble with row ID, original ID, prompt word, and dynamic forward flow values.
}
\description{
This function calculates the forward flow of a list of words per Gray et. al, 2019. Forward flow is a way to quantify the forward motion of naturalistic thought. Research suggests that it is correlated with creativity.  See forwardflow.org for more information.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com> & Brendan Baker <bsb5477@psu.edu>
}
