% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/network_backbone_extraction.r
\name{getBackboneNetwork}
\alias{getBackboneNetwork}
\title{Extract the backbone of a network.}
\usage{
getBackboneNetwork(g, alpha = 0.05, direction = "none",
  delete.isolates = T, max.vertices = NULL)
}
\arguments{
\item{g}{A graph in the `Igraph` format.}

\item{alpha}{The threshold for the alpha. Can be interpreted similar to a p value (see paper for clarrification).}

\item{direction}{direction = 'none' can be used for both directed and undirected networks, and is (supposed to be) the disparity filter proposed in Serrano et al. (2009) is used. By setting to 'in' or 'out', the alpha is only calculated for out or in edges. This is an experimental use of the backbone extraction (so beware!) but it seems a logical application.}

\item{delete.isolates}{If TRUE, vertices with degree 0 (i.e. no edges) are deleted.}

\item{max.vertices}{Optional. Set a maximum number of vertices for the network to be produced. The alpha is then automatically lowered to the point that only the given number of vertices remains connected (degree > 0). This can be usefull if the purpose is to make an interpretation friendly network. See e.g., http://jcom.sissa.it/archive/14/01/JCOM_1401_2015_A01}
}
\value{
A graph in the Igraph format
}
\description{
Based on the following paper: Serrano, M. Á., Boguñá, M., & Vespignani, A. (2009). Extracting the multiscale backbone of complex weighted networks. Proceedings of the National Academy of Sciences, 106(16), 6483-6488.
}

