% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_adjacency.r
\name{getQueryMatrix}
\alias{getQueryMatrix}
\title{Creates a matrix where rows match the tokens, columns represent the queries in query_regex, and values represent the word distance to each query hit}
\usage{
getQueryMatrix(tokens, query_regex, doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"),
  word.col = getOption("word.col", "word"), presorted = F,
  default.window = NA, return_i = NULL)
}
\arguments{
\item{default.window}{}
}
\value{
a sparse matrix
}
\description{
Creates a matrix where rows match the tokens, columns represent the queries in query_regex, and values represent the word distance to each query hit
}

