% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{mergeMatrices}
\alias{mergeMatrices}
\title{Merge a list of matrices}
\usage{
mergeMatrices(matrices, reduce_func = "+")
}
\arguments{
\item{matrices}{A list of matrices}

\item{reduce_func}{The function to pass to Reduce}
}
\value{
A single matrix
}
\description{
Merge matrices, by standradizing column and row dimensions and using the Reduce function to merge them.
}

