% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experimental.r
\name{mergeTermClusters}
\alias{mergeTermClusters}
\title{Merge terms with a high mutual direction conditional probability}
\usage{
mergeTermClusters(m, min.similarity = 0.95, max.label_length = 3)
}
\arguments{
\item{m}{A sparse matrix in which columns are terms. Can be a DocumentTermMatrix class from the tm package}

\item{min.similarity}{The minimum conditional probability. The conditional probability of two terms in both directions needs to be higher than min.similarity for terms to be merged}

\item{max.label_length}{Terms that are merged together will be collapsed into a single label. To prevent very long labels, this is cut of from the [max.label_length] term.}
}
\value{
a matrix (or document term matrix)
}
\description{
Merge terms that are likely to occur together. Specifically, terms of which the conditinal probability is higher than min.similarity in both directions. Or: P(A|B) >= min.similarity & P(B|A) >= min.similarity)
Not that this is not always a good thing. Unrelated terms that always occur together will also be merged. Whether this makes sense depends on the type of analysis.
}

