% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.high.R
\docType{data}
\name{net.high}
\alias{net.high}
\title{High Openness to Experience Network}
\format{net.high (matrix, 160 x 160)}
\usage{
data(net.high)
}
\description{
High openness to experience network from Christensen & Kenett (2019)
}
\examples{
data("net.high")
}
\references{
Christensen, A. P., & Kenett, Y. N. (2019)
Semantic network analysis (SemNA): A tutorial on preprocessing, estimating, and analyzing semantic networks.
\emph{PsyArXiv}.
doi:\href{https://doi.org/10.31234/osf.io/eht87}{10.31234/osf.io/eht87}
}
\keyword{datasets}
