% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.low.R
\docType{data}
\name{net.low}
\alias{net.low}
\title{Low Openness to Experience Network}
\format{net.low (matrix, 160 x 160)}
\usage{
data(net.low)
}
\description{
Low openness to experience network from Christensen & Kenett (2019)
}
\examples{
data("net.low")
}
\references{
Christensen, A. P., & Kenett, Y. N. (2019)
Semantic network analysis (SemNA): A tutorial on preprocessing, estimating, and analyzing semantic networks.
\emph{PsyArXiv}.
doi:\href{https://doi.org/10.31234/osf.io/eht87}{10.31234/osf.io/eht87}
}
\keyword{datasets}
