% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRW.R
\name{NRW}
\alias{NRW}
\title{Naive Random Walk Network Estimation}
\usage{
NRW(data, threshold = 3)
}
\arguments{
\item{data}{Matrix or data frame.
A preprocessed verbal fluency matrix where
rows are participants and columns are verbal fluency
responses}

\item{threshold}{Numeric.
Minimum number of total co-occurrences to be included
in the network}
}
\value{
Returns a undirected semantic network
}
\description{
Estimates a semantic network using the Naive Random Walk
method described in Lerner, Ogrocki, and Thomas (2009)
}
\examples{
# Get data
data <- SemNetCleaner::bad.response(as.matrix(SemNetCleaner::open.animals[,-c(1:3)]), 99)

# Organize group data
## Get group data
group <- SemNetCleaner::open.animals$Group

## Low and high openness to experience groups
low <- data[which(group == 1),]
high <- data[which(group == 2),]

# Compute networks
low.net <- NRW(low)
high.net <- NRW(high)

}
\references{
Lerner, A. J., Ogrocki, P. K., & Thomas, P. J. (2009).
Network graph analysis of category fluency testing.
\emph{Cognitive and Behavioral Neurology}, \emph{22}, 45-52.
\href{https://doi.org/10.1097/WNN.0b013e318192ccaf}{https://doi.org/10.1097/WNN.0b013e318192ccaf}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
