% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.result.R
\docType{data}
\name{one.result}
\alias{one.result}
\title{Simulated Result for Dataset One}
\format{
one.result (list, length = 4)
}
\usage{
data(one.result)
}
\description{
A result of \code{\link[SemNeT]{bootSemNeT}} from a simulated dataset
}
\examples{
data("one.result")
}
\keyword{datasets}
