% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.binary.R
\docType{data}
\name{open.binary}
\alias{open.binary}
\title{Binary response Matrices (Openness and Verbal Fluency)}
\format{
open.binary (matrix, 516 x 367)
}
\usage{
data(open.binary)
}
\description{
Binary response matrices for the Animals verbal fluency data (\emph{n} = 516)
from Christensen et al. (2018).
}
\examples{
data("open.binary")
}
\references{
Christensen, A. P., Kenett, Y. N., Cotter, K. N., Beaty, R. E., & Silvia, P. J. (2018).
Remotely close associations: Openness to experience and semantic memory structure.
\emph{European Journal of Personality}, \emph{32}, 480-492.
doi:\href{https://doi.org/10.1002/per.2157}{10.1002/per.2157}
}
\keyword{datasets}
