% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.clean.R
\docType{data}
\name{open.clean}
\alias{open.clean}
\title{Cleaned response Matrices (Openness and Verbal Fluency)}
\format{
open.clean (matrix, 516 x 35)
}
\usage{
data(open.clean)
}
\description{
Cleaned response matrices for the Animals verbal fluency data (\emph{n} = 516)
from Christensen et al. (2018).
}
\examples{
data("open.clean")
}
\references{
Christensen, A. P., Kenett, Y. N., Cotter, K. N., Beaty, R. E., & Silvia, P. J. (2018).
Remotely close associations: Openness to experience and semantic memory structure.
\emph{European Journal of Personality}, \emph{32}, 480-492.
doi:\href{https://doi.org/10.1002/per.2157}{10.1002/per.2157}
}
\keyword{datasets}
