% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.group.R
\docType{data}
\name{open.group}
\alias{open.group}
\title{Groups for Openness and Verbal Fluency}
\format{
open.group (vector, length = 516)
}
\usage{
data(open.group)
}
\description{
Groups for the Animals verbal fluency data (\emph{n} = 516)
from Christensen et al. (2018; see also \code{\link[SemNeT]{open.clean}}).
}
\examples{
data("open.group")
}
\references{
Christensen, A. P., Kenett, Y. N., Cotter, K. N., Beaty, R. E., & Silvia, P. J. (2018).
Remotely close associations: Openness to experience and semantic memory structure.
\emph{European Journal of Personality}, \emph{32}, 480-492.
doi:\href{https://doi.org/10.1002/per.2157}{10.1002/per.2157}
}
\keyword{datasets}
