% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.plot.R
\name{org.plot}
\alias{org.plot}
\title{Organization function for partboot.plot}
\usage{
org.plot(bootData, bootPaired, sampsData, sampsPaired, len, measname,
  netmeas, pall, paired, CI, labels)
}
\arguments{
\item{bootData}{bootData from \link[SemNeT]{partboot}}

\item{bootPaired}{bootPaired from \link[SemNeT]{partboot}}

\item{sampsData}{sampsData from \link[SemNeT]{partboot}}

\item{sampsPaired}{sampsPaired from \link[SemNeT]{partboot}}

\item{len}{Number of samples in data list}

\item{measname}{Full network measure name}

\item{netmeas}{Abbreviated network measure name}

\item{pall}{Color palette to be used from \code{\link{RColorBrewer}}}

\item{paired}{Is samples paired?}

\item{CI}{Confidence interval to be used}

\item{labels}{Labels to be used in plot.}
}
\value{
Returns plots for the specified measures
}
\description{
A wrapper function used in \link[SemNeT]{partboot}. Not to be used individually
}
\examples{
###NOT FOR INDIVIDUAL USE###
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
