% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.partboot.R
\name{plot.partboot}
\alias{plot.partboot}
\title{Plot for partboot}
\usage{
\method{plot}{partboot}(..., groups = NULL, measures = c("ASPL", "CC",
  "Q"))
}
\arguments{
\item{...}{Object(s) from \code{\link[SemNeT]{partboot}}}

\item{groups}{Character.
Labels for groups in the order they were entered
in \code{\link[SemNeT]{partboot}} (e.g.,
\code{data} = first,
\code{paired} = second)}

\item{measures}{Character.
Measures to be plotted}
}
\value{
Returns plots for the specified measures
}
\description{
Bootstraps (without replacement) the nodes in the network and computes global network characteristics
}
\examples{
# Finalize rmatA
finalCmat <- SemNetCleaner::finalize(SemNetCleaner::convmat)
# Finalize rmatB
finalRmat <- SemNetCleaner::finalize(SemNetCleaner::rmat)

# Equate rmatA and rmatB
eq1 <- SemNetCleaner::equate(finalCmat,finalRmat)

# Obtain respective equated response matrices
eqCmat <- eq1$rmatA
eqRmat <- eq1$rmatB

\dontrun{

# Run partial bootstrap networks
results <- partboot(data = eqCmat, paired = eqRmat,
percent = .50, iter = 1000, sim = "cosine", cores = 4)

# Plot
plot(results, groups = c("eqCmat","eqRmat"))

}

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
