% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SemNeT.R
\name{plot.SemNeT}
\alias{plot.SemNeT}
\title{Plots a Single Network}
\usage{
\method{plot}{SemNeT}(
  A,
  weighted = FALSE,
  node_size = 7,
  node_color = "#EFB0A1",
  labels = FALSE,
  label_size = 4.5,
  edge_size = 1,
  edge_color = "grey",
  positive_edge_color = "green",
  negative_edge_color = "red",
  layout = c("qgraph_spring", "eigen", "fruchtermanreingold", "geodist", "mds",
    "princoord", "spring", "springrepulse"),
  layout_args = list(),
  expand_plot = 0,
  seed = 1234,
  produce = TRUE,
  ...
)
}
\arguments{
\item{A}{Matrix or data frame.
An adjacency matrix of a network}

\item{weighted}{Boolean.
Should networks be plotted with weights?
Defaults to \code{FALSE}.
Set to \code{TRUE} to plot networks with weights
corresponding to association strength}

\item{node_size}{Numeric.
Size of the nodes in the network.
Defaults to \code{7}}

\item{node_color}{Character or hex code.
Color of the nodes in the network.
Defaults to \code{"#EFB0A1"}}

\item{labels}{Boolean.
Should node labels be used?
Defaults to \code{FALSE}.
Labels may add noise and distract from understanding
the broader network structures.
Set to \code{TRUE} to plot labels.
Use other label arguments to get them the way you want}

\item{label_size}{Numeric.
Size of labels.
Defaults to \code{4.5}}

\item{edge_size}{Numeric.
Size of edges in the network.
Defaults to \code{1}}

\item{edge_color}{Character or hex code.
Color of the edges in the network.
Defaults to \code{"grey"}}

\item{positive_edge_color}{Character or hex code.
If \code{weighted = TRUE}, then positive and negative
edges will be different colors. This argument sets
the color of the positively signed edges.
Defaults to \code{"green"}}

\item{negative_edge_color}{Character or hex code.
If \code{weighted = TRUE}, then positive and negative
edges will be different colors. This argument sets
the color of the negatively signed edges.
Defaults to \code{"red"}}

\item{layout}{Character.
Changes the layout of the nodes in the network.
Defaults to \code{"qgraph_spring"}, which uses
a version of the Fruchterman-Reingold algorithm 
(see \code{\link[qgraph]{qgraph.layout.fruchtermanreingold}}).
Other options are taken from \code{\link[sna]{gplot.layout}}}

\item{layout_args}{List.
Takes in arguments for \code{layout}.
See \code{\link[sna]{gplot.layout}} for more information}

\item{expand_plot}{Positive numbers.
Sometimes plots will cut of node labels. This argument
can "expand" the plot so that node labels are not cut-off.
Defaults to \code{0} or no adjustment.
Use positive numbers to increase the space available for the
network to be plot}

\item{seed}{Integer.
Sets seed to have reproducible layouts.
Defaults to \code{1234}.
Use different numbers to switch up the layout
(might not change some layouts)}

\item{produce}{Boolean.
Should plot be produced?
Defaults to \code{TRUE}.
Used mainly for other functions (e.g., \code{\link{compare_nets}})}

\item{...}{Additional arguments for \code{\link[GGally]{ggnet2}}}
}
\value{
Plots networks using \code{\link[GGally]{ggnet2}}
}
\description{
Uses \code{\link[GGally]{ggnet2}} to plot networks
}
\examples{
# Simulate Datasets
one <- sim.fluency(10)

# Compute similarity matrix
cos1 <- similarity(one, method = "cosine")

# Compute networks
net1 <- TMFG(cos1)

# Compare networks
plot(net1)

# Change node size
plot(net1, node_size = 14)

}
\references{
Epskamp, S., Cramer, A. O. J., Waldorp, L. J., Schmittmann, V. D., & Borsboom, D. (2012).
qgraph: Network visualizations of relationships in psychometric data.
\emph{Journal of Statistical Software}, \emph{48}, 1-18.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
