% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randwalk.R
\name{randwalk}
\alias{randwalk}
\title{Random Walk Simulation}
\usage{
randwalk(A, B, reps = 20, steps = 10, iter = 10000)
}
\arguments{
\item{A}{Adjacency matrix of a semantic network}

\item{B}{A comparison adjacency matrix of a semantic network}

\item{reps}{Number of repetitions of increments in 10 steps.
Defaults to 20}

\item{steps}{Number of random steps to begin with.
Defaults to 10}

\item{iter}{Number of iterations for each random walk}
}
\value{
A result matrix containing the means and standard deviations for
several measures as well as p-values for a Mann-Whitney U test
}
\description{
Simulates random walks over two networks to examine the characteristics
of spontaneous spreading activation
}
\examples{
#finalize rmatA
finalCmat <- finalize(convmat)
#finalize rmatB
finalRmat <- finalize(rmat)

#equate rmatA and rmatB
eq1 <- equate(finalCmat,finalRmat)

#obtain respective equated response matrices
eqCmat <- eq1$rmatA
eqRmat <- eq1$rmatB

\donttest{
rw.results <- randwalk(eqCmat,eqRmat)
}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
Yoed Kenett <yoedkenett@gmail.com>
}
