% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experimental.r
\name{reduceLabelOverlap}
\alias{reduceLabelOverlap}
\title{Reduce overlap of labels in igraph network}
\usage{
reduceLabelOverlap(g, fontsize_multiplier = 5)
}
\arguments{
\item{g}{A graph in the Igraph format, which has the g$layout attribute (set by using the layout. functions)}

\item{fontsize_multiplier}{The wordlayout function takes fontsize into account to determine overlap. As a temporary solution, increasing font size decreased label overlap (but messes up the network if done extremely)}
}
\value{
a network layout, to be assigned to g$layout
}
\description{
This function does not yet work properly. The idea is to use the wordlayout function of the wordcloud package to arrange nodes in order to prevent label overlap.
}

