% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_query.r
\name{reportHitcount}
\alias{reportHitcount}
\title{Title}
\usage{
reportHitcount(tokens, hits = NULL, token_i = NULL,
  doc.col = getOption("doc.col", "doc.id"))
}
\arguments{
\item{hits}{}
}
\description{
Title
}

