% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_query.r
\name{reportKWIC}
\alias{reportKWIC}
\title{Title}
\usage{
reportKWIC(tokens, hits = NULL, token_i = NULL, nwords = 10,
  nsample = 10, random_sample = T, doc.col = getOption("doc.col",
  "doc_id"), position.col = getOption("position.col", "position"),
  word.col = getOption("wor.col", "word"))
}
\arguments{
\item{word.col}{}
}
\description{
Title
}

