% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_query.r
\name{reportSummary}
\alias{reportSummary}
\title{Title}
\usage{
reportSummary(tokens, hits = NULL, token_i = NULL, hitcount = T,
  tokenfreq = T, keywordIC = T, kwic_nwords = 10, kwic_sample = 10,
  random_sample = T, doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"),
  word.col = getOption("wor.col", "word"))
}
\arguments{
\item{hitcount}{}
}
\description{
Title
}

