% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_query.r
\name{reportTokenFreq}
\alias{reportTokenFreq}
\title{Title}
\usage{
reportTokenFreq(hits, token_i = NULL, doc.col = getOption("doc.col",
  "doc.col"), word.col = get.Option("word", "word"))
}
\arguments{
\item{word.col}{}
}
\description{
Title
}

