% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmat.R
\docType{data}
\name{rmat}
\alias{rmat}
\title{Cleaned trial data of verbal fluency responses}
\format{A 15x95 response matrix}
\usage{
data(rmat)
}
\description{
Cleaned trial data of animal verbal fluency responses. The columns are responses and the
rows are the participants.
}
\examples{

data(rmat)
}
\keyword{datasets}
