% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticNetworkCleaner.R
\name{semnetboot}
\alias{semnetboot}
\title{Partial Bootstrapped Semantic Network Analysis}
\usage{
semnetboot(data, nodes, iter = 1000, corr = c("cor", "cosine"), cores,
  seeds = NULL)
}
\arguments{
\item{data}{Cleaned response matrix}

\item{nodes}{Number of nodes for partial bootstrap.
Defaults to half}

\item{iter}{Number of iterations in bootstrap.
Defaults to 1000}

\item{corr}{Association method to use.
Defaults to "cosine"}

\item{cores}{Number of computer processing cores to use for bootstrapping samples.
Defaults to \emph{n} - 1 total number of cores.
Set to any number between 1 and maxmimum amount of cores on your computer}

\item{seeds}{Seeds used in previous run.
Defaults to NULL.
Input a vector from previous run to replicate analyses}
}
\value{
Returns a list that includes the original semantic network measures (origmeas; ASPL, CC, Q, S),
the bootstrapped semantic network measures (bootmeas),
and Seeds that can be used to replicate analysis
}
\description{
Bootstraps (without replacement) the nodes in the network and computes global network characteristics
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
