% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semantic Network Cleaner.R
\name{semnetcleaner}
\alias{semnetcleaner}
\title{Semantic Network Cleaner}
\usage{
semnetcleaner(data)
}
\arguments{
\item{data}{A dataset of verbal fluency or linguistic data (read in data "as.is")}
}
\value{
A binary matrix of responses (rows = participants, cols = responses)
}
\description{
An automated cleaning function for semantic network data
}
\examples{
\dontrun{data<-read.csv(file.choose(),header=FALSE,sep=",",as.is=TRUE)}
\dontrun{responsematrix<-semnetcleaner(data)}
}
\references{
Hornik, K., & Murdoch, D. (2010).
Watch Your Spelling!.
The R Journal, 3(2), 22-28.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
