% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semnetmeas.R
\name{semnetmeas}
\alias{semnetmeas}
\title{Semantic Network Measures}
\usage{
semnetmeas(A, weighted = FALSE)
}
\arguments{
\item{A}{Matrix or data frame.
An adjacency matrix of a network}

\item{weighted}{Boolean.
Should weighted measures be computed?
Defaults to \code{FALSE}.
Set to \code{TRUE} for weighted measures}
}
\value{
Returns a values for ASPL, CC, and Q
}
\description{
Computes the average shortest path length (ASPL),
clustering coefficient(CC), and modularity (Q) of the network
}
\examples{
# Finalize convmat
fin <- SemNetCleaner::finalize(SemNetCleaner::convmat)

# Compute cosine similarity
cosL <- similarity(fin, method = "cosine")

# Construct network
A <- NetworkToolbox::TMFG(cosL)$A

# Compute global network measures
globmeas <- semnetmeas(A)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
