% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemNeTShiny.R
\name{SemNeTShiny}
\alias{SemNeTShiny}
\title{Shiny App for \code{\link{SemNeT}}}
\usage{
SemNeTShiny()
}
\value{
A list called \code{resultShiny} containing:

\item{data}{The data imported into \code{\link[SemNeT]{SemNetShiny}}}

\item{group}{The grouping variable imported into \code{\link[SemNeT]{SemNetShiny}}}

\item{network}{The networks generated during \code{\link[SemNeT]{SemNetShiny}} session.
The networks are labelled using the provided grouping variable}

\item{measures}{Network measures ASPL (Average Shortest Path Lengths),
CC (Clustering Coefficient), and Q (Modularity) for the networks generated
during \code{\link[SemNeT]{SemNetShiny}}}

\item{comparePlot}{A visualization of the networks generated during \code{\link[SemNeT]{SemNetShiny}}}

\item{randomTest}{Statistical results from the Random Network Test in \code{\link[SemNeT]{SemNetShiny}}
(see \code{\link[SemNeT]{randnet.test}})}

\item{bootstrap}{Results from the Bootstrap Network Analysis in \code{\link[SemNeT]{SemNetShiny}}
(see \code{\link[SemNeT]{bootSemNeT}})}

\item{bootstrapTest}{Statistical results from the Bootstrap Network Analysis
(see \code{\link[SemNeT]{test.bootSemNeT}})}

\item{bootstrapPlot}{Plots of the statistical results from the Bootstrap Network Analysis
(see \code{\link[SemNeT]{plot.bootSemNeT}})}
}
\description{
An interactive Shiny application for running \code{\link{SemNeT}} analysis.
}
\examples{

if(interactive())
{SemNeTShiny()}


}
