% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.r
\name{setTokenlistColnames}
\alias{setTokenlistColnames}
\title{Set which columns in the tokens data frame should be used}
\usage{
setTokenlistColnames(doc.col = "doc_id", position.col = "position",
  word.col = "word")
}
\arguments{
\item{doc.col}{The name of the column containing the unique id's for documents}

\item{position.col}{The name of the column contaning the positions of a token in a document}

\item{word.col}{The name of the column contaning the token text (e.g., word, lemma)}
}
\value{
Nothing. Sets the names globally (in options())
}
\description{
Set which columns in the tokens data frame should be used
}

