% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fluency.R
\name{sim.fluency}
\alias{sim.fluency}
\title{Convert Adjacency Matrix to Cytoscape Format}
\usage{
sim.fluency(nodes)
}
\arguments{
\item{nodes}{Numeric.
Number of nodes to simulate in data}
}
\value{
A binary matrix with \code{p} (participants) by \code{n} (nodes)
}
\description{
Simulates verbal fluency data based on the number of
nodes in the desired network. Participants are 5% greater than
the maximum simulated fluency response given
}
\details{
Simulates data
}
\examples{
# Simulate data for 50 nodes
sim.fluency(nodes = 50)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
