% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{simple_dtm}
\alias{simple_dtm}
\title{Simple dtm to illustrate coOccurenceNetwork function}
\format{simple_dtm: A small document term matrix}
\usage{
data(simple_dtm)
}
\description{
Simple dtm to illustrate coOccurenceNetwork function
}
\keyword{datasets}

