% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sotu_tokenlist}
\alias{sotu_tokenlist}
\title{Tokens from Bush and Obama's State of the Union addresses}
\format{sotu.tokens: A dataframe of tokens produced by using Stanford CoreNLP to lemmatize the State of the Union speeches; and sotu.meta the metadata for these documents}
\usage{
data(sotu_tokenlist)
}
\description{
Tokens from Bush and Obama's State of the Union addresses
}
\keyword{datasets}

