% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenlist.r
\name{tokenlistToDTM}
\alias{tokenlistToDTM}
\title{Transform a tokenlist into a document-term matrix (DTM)}
\usage{
tokenlistToDTM(tokenlist, doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"),
  word.col = getOption("word.col", "word"))
}
\arguments{
\item{tokenlist}{}

\item{doc.col}{}

\item{position.col}{}

\item{word.col}{}
}
\value{
a DTM, as a sparse matrix in the dgTMatrix class
}
\description{
Transform a tokenlist into a document-term matrix (DTM)
}

