% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\docType{data}
\name{trial}
\alias{trial}
\title{Trial data of verbal fluency responses}
\format{A 49x15 response matrix}
\usage{
data(trial)
}
\description{
Trial data of animal verbal fluency responses. The columns are participants and the
rows are their responses.
}
\examples{

data(trial)
}
\keyword{datasets}
