% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.result.R
\docType{data}
\name{two.result}
\alias{two.result}
\title{Simulated Result for Dataset One and Two}
\format{
two.result (list, length = 6)
}
\usage{
data(two.result)
}
\description{
A result of \code{\link[SemNeT]{bootSemNeT}} from two simulated datasets
}
\examples{
data("two.result")
}
\keyword{datasets}
