% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.plots.R
\docType{data}
\name{vignette.plots}
\alias{vignette.plots}
\title{Plots for Vignette}
\format{vignette.plots (list, length = 3)}
\usage{
data(vignette.plots)
}
\description{
Plots for vignette taken from Christensen & Kenett (2019)
}
\examples{
data("vignette.plots")
}
\references{
Christensen, A. P., & Kenett, Y. N. (2019)
Semantic network analysis (SemNA): A tutorial on preprocessing, estimating, and analyzing semantic networks.
\emph{PsyArXiv}.
doi:\href{https://doi.org/10.31234/osf.io/eht87}{10.31234/osf.io/eht87}
}
\keyword{datasets}
