% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{windowedCoOccurenceNetwork}
\alias{windowedCoOccurenceNetwork}
\title{A sliding window approach to calculate the co-occurence of words}
\usage{
windowedCoOccurenceNetwork(location, term, context, window.size = 3,
  output.per.context = F, two.sided = T)
}
\arguments{
\item{location}{An integer vector giving the position of terms in a given context (e.g., document, paragraph, sentence)}

\item{term}{A character vector giving the terms}

\item{context}{A vector giving the context in which terms occur (e.g., document, paragraph, sentence)}

\item{window.size}{The distance within which words should occur from each other to be counted as a co-occurence.}

\item{output.per.context}{Logical. If True, co-occurences are reported per context (beware that this takes longer and can lead to huge output)}

\item{two.sided}{Logical. If false, it is only counted how often a word occured `after` another word within the given window size}
}
\value{
An edgelist (data.frame) with columns x, y and weight, in which weight represents the number of times y occured within a [window.size] word distance from x. If output.per.context is True, co-occurences are reported per context, and the edgelist has an additional context column.
}
\description{
A sliding window approach to calculate the co-occurence of words
}

