% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wordWindowOccurence}
\alias{wordWindowOccurence}
\title{Gives the window in which a term occured in a matrix.}
\usage{
wordWindowOccurence(location, term, context, window.size = 3, two.sided = T)
}
\arguments{
\item{location}{An integer vector giving the position of terms in a given context (e.g., document, paragraph, sentence)}

\item{term}{A character vector giving the terms}

\item{context}{A vector giving the context in which terms occur (e.g., document, paragraph, sentence)}

\item{window.size}{The distance within which words should occur from each other to be counted as a co-occurence.}

\item{two.sided}{Logical. If false, it is only counted how often a word occured `after` another word within the given window size}
}
\value{
A list with two matrices. location.mat gives the specific location of a term, and window.mat gives the window in which each word occured. The rows represent the location of a term, and matches the input of this function (location, term and context). The columns represents terms.
}
\description{
This function returns the occurence of words (location.matrix) and the window of occurence (window.matrix). This format enables the co-occurence of words within sliding windows (i.e. word distance) to be calculated by multiplying location.matrix with window.matrix.
}

