% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_query.r
\name{YnotX}
\alias{YnotX}
\title{Title}
\usage{
YnotX(hits.x, hits.y, doc.col = getOption("doc.col", "doc_id"),
  position.col = getOption("position.col", "position"))
}
\arguments{
\item{...}{}
}
\description{
Title
}

