% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converge.R
\name{converge}
\alias{converge}
\title{Converge Responses}
\usage{
converge(rmat, word, replace)
}
\arguments{
\item{rmat}{A semnetcleaner filtered response matrix}

\item{word}{The column name that will incoporate the \strong{replace} column's binarized responses (must be characters)}

\item{replace}{The column name that should be merged with the \strong{word} column (must be characters)}
}
\value{
The response matrix with the \strong{word} column merged and the \strong{replace} column removed
}
\description{
Merge a column of binarized response data with another
}
\examples{
#converge "kitten" into response of "cat"
rmat <- converge(rmat,"cat","kitten")
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
