% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2cytoscape.R
\name{convert2cytoscape}
\alias{convert2cytoscape}
\title{Convert Adjacency Matrix to Cytoscape Format}
\usage{
convert2cytoscape(A)
}
\arguments{
\item{A}{Matrix or data frame.
A cleaned, finalized response matrix ready to be visualized}
}
\value{
A sparse matrix formatted for Cytoscape
}
\description{
Converts an adjacency matrix to Cytoscape's sparse matrix format
}
\examples{
# Finalize rmatA
finalCmat <- SemNetCleaner::finalize(SemNetCleaner::convmat)
# Finalize rmatB
finalRmat <- SemNetCleaner::finalize(SemNetCleaner::rmat)

# Equate rmatA and rmatB
eq1 <- SemNetCleaner::equate(finalCmat,finalRmat)

# Obtain respective equated response matrices
eqCmat <- eq1$rmatA
eqRmat <- eq1$rmatB

# Compute similarity matrix
cosC <- similarity(eqCmat, method = "cosine")
cosR <- similarity(eqRmat, method = "cosine")

# Compute networks using NetworkToolbox
Cnet <- NetworkToolbox::TMFG(cosC)$A
Rnet <- NetworkToolbox::TMFG(cosR)$A

# Cnovert to Cytoscape format
ctyoC <- convert2cytoscape(Cnet)
ctyoR <- convert2cytoscape(Rnet)

\dontrun{

# Export to .csv
write.csv(cytoC, "netC.csv", row.names = FALSE)
write.csv(cytoR, "netR.csv", row.names = FALSE)

}

}
\references{
Shannon, P., Markiel, A., Ozier, O., Baliga, N. S., Wang, J. T., Ramage, D., ... & Ideker, T. (2003).
Cytoscape: A software environment for integrated models of biomolecular interaction networks.
\emph{Genome Research}, \emph{13}, 2498-2504.
doi:\href{http://www.genome.org/cgi/doi/10.1101/gr.1239303}{10.1101/gr.1239303}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
