% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{coOccurenceNetwork}
\alias{coOccurenceNetwork}
\title{Create an adjacency graph from a document term matrix}
\usage{
coOccurenceNetwork(dtm, measure = "cooccurence")
}
\arguments{
\item{dtm}{a document term matrix, either or not in the dtm format from the `tm` package}

\item{measure}{the measure to calcualte adjacency. Currently supports cosine and conditional probability}
}
\value{
A graph in the Igraph format in which edges represent the adjacency of terms
}
\description{
Create an adjacency graph from a document term matrix
}

