% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semantic Network Cleaner.R
\name{finalize}
\alias{finalize}
\title{Finalize Function}
\usage{
finalize(rmat)
}
\arguments{
\item{rmat}{A semnetcleaner and converge filtered response matrix}
}
\value{
A matrix with responses given by two or more people
}
\description{
Finalizes the response matrix by keeping responses that are given by two or more people
}
\examples{
finalRmat <- finalize(rmat)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
