% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.plot.R
\name{org.plot}
\alias{org.plot}
\title{Organization function for partboot.plot}
\usage{
org.plot(input, paired, len, measures, groups, netmeas)
}
\arguments{
\item{input}{List.
Input data from \link[SemNeT]{partboot}}

\item{paired}{Numeric.
\code{1} equals single sample.
\code{2} equals paired samples}

\item{len}{Numeric.
Number of bootstrapped samples (percentages)}

\item{measures}{Character.
Network measures to be entered}

\item{groups}{Character.
Names for the group(s)}

\item{netmeas}{Character.
Abbreviated network measure name that should be plotted}
}
\value{
Returns plots for the specified measures
}
\description{
A wrapper function used in \link[SemNeT]{partboot}. Not to be used individually
}
\examples{
#### NOT INTENDED FOR INDIVIDUAL USE ####
#### WRAPPER FUNCTION ####

# Finalize rmatA
finalCmat <- SemNetCleaner::finalize(SemNetCleaner::convmat)
# Finalize rmatB
finalRmat <- SemNetCleaner::finalize(SemNetCleaner::rmat)

# Equate rmatA and rmatB
eq1 <- SemNetCleaner::equate(finalCmat,finalRmat)

# Obtain respective equated response matrices
eqCmat <- eq1$rmatA
eqRmat <- eq1$rmatB

\dontrun{

# Run partial bootstrap networks
results <- partboot(data = eqCmat, paired = eqRmat,
percent = .50, iter = 1000, sim = "cosine", cores = 4)

# Plot
plot(results, groups = c("eqCmat","eqRmat"))

}

}
\references{
Allen, M., Poggiali, D., Whitaker, K., Marshall, T. R., Kievit, R. (2018).
Raincloud plots: A multi-platform tool for robust data visualization.
\emph{PeerJ Preprints}, \emph{6}, e27137v1.
\href{https://doi.org/10.7287/peerj.preprints.27137v1}{10.7287/peerj.preprints.27137v1}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
