% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticNetworkCleaner.R
\name{partboot.plot}
\alias{partboot.plot}
\title{Partial Bootstrapped Semantic Network Analysis Plot}
\usage{
partboot.plot(object, paired = FALSE, CI = 0.975, labels = NULL,
  measures = c("ASPL", "CC", "Q", "S", "TotalStrength"))
}
\arguments{
\item{object}{An object from \link[SemNetCleaner]{partboot}}

\item{paired}{Is object from a paired \link[SemNetCleaner]{partboot}?}

\item{CI}{Confidence intervals to use for plot.
Defaults to .975}

\item{labels}{Labels to be used in plot.
Defaults to NULL.
Typed responses will be requested if NULL}

\item{measures}{Measures to be plotted}
}
\value{
Returns plots for the specified measures
}
\description{
Bootstraps (without replacement) the nodes in the network and computes global network characteristics
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
