% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticNetworkCleaner.R
\name{semnetcleaner}
\alias{semnetcleaner}
\title{Semantic Network Cleaner}
\usage{
semnetcleaner(data, miss = 99)
}
\arguments{
\item{data}{A dataset of verbal fluency or linguistic data}

\item{miss}{Value for missing data. Defaults to 99}
}
\value{
A list of a binary matrix of responses (binary; rows = participants, columns = responses) and cleaned response matrix (responses)
}
\description{
An automated cleaning function for semantic network data
}
\examples{
\dontrun{
rmat<-semnetcleaner(trial)
}
}
\references{
Hornik, K., & Murdoch, D. (2010).
Watch Your Spelling!.
\emph{The R Journal}, 3(2), 22-28.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
