% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemanticNetworkCleaner.R
\name{semnetcleaner}
\alias{semnetcleaner}
\title{Semantic Network Cleaner}
\usage{
semnetcleaner(data)
}
\arguments{
\item{data}{A dataset of verbal fluency or linguistic data (rows = responses, columns = participants)}
}
\value{
A binary matrix of responses (rows = participants, columns = responses) and unique responses of the response matrix
}
\description{
An automated cleaning function for semantic network data
}
\examples{
\dontrun{

data<-read.csv(file.choose(),header=FALSE,sep=",",as.is=TRUE)

rmat<-semnetcleaner(trial)
}
}
\references{
Hornik, K., & Murdoch, D. (2010).
Watch Your Spelling!.
\emph{The R Journal}, 3(2), 22-28.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
