% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemNeTShiny.R
\name{SemNeTShiny}
\alias{SemNeTShiny}
\title{SemNeTShiny}
\usage{
SemNeTShiny()
}
\value{
Nothing
}
\description{
An interactive Shiny application for running \code{\link[SemNeT]{SemNeT-package}} analysis.
}
\details{
Start SemNeTShiny

This starts the Shiny application for \code{\link[SemNeT]{SemNeT-package}}
}
\examples{
\dontrun{
SemNeTShiny()
}
}
\keyword{CTT}
