% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_adjacency.r
\name{windowedCoOccurenceNetwork}
\alias{windowedCoOccurenceNetwork}
\title{A sliding window approach to calculate the co-occurence of words}
\usage{
windowedCoOccurenceNetwork(tokenlist, window.size = 10,
  output.per.context = F, direction = "<>", doc.col = getOption("doc.col",
  "doc_id"), position.col = getOption("position.col", "position"),
  word.col = getOption("word.col", "word"))
}
\arguments{
\item{tokenlist}{}

\item{window.size}{The distance within which words should occur from each other to be counted as a co-occurence.}

\item{output.per.context}{Logical. If True, co-occurences are reported per context (beware that this takes longer and can lead to huge output)}

\item{direction}{a string indicating whether only the left ('<') or right ('>') side of the window, or both ('<>'), should be used.}
}
\value{
An edgelist (data.frame) with columns x, y and weight, in which weight represents the number of times y occured within a [window.size] word distance from x. If output.per.context is True, co-occurences are reported per context, and the edgelist has an additional context column.
}
\description{
A sliding window approach to calculate the co-occurence of words
}

