% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjacency_functions.r
\name{wordWindowAdjacency}
\alias{wordWindowAdjacency}
\title{A sliding window approach to calculate the co-occurence of words}
\usage{
wordWindowAdjacency(location, term, context, window.size = 3, two.sided = T,
  count.once = T)
}
\arguments{
\item{location}{An integer vector giving the position of terms in a given context (e.g., document, paragraph, sentence)}

\item{term}{A character vector giving the terms}

\item{context}{A vector giving the context in which terms occur (e.g., document, paragraph, sentence)}

\item{window.size}{The distance within which words should occur from each other to be counted as a co-occurence.}

\item{two.sided}{Logical. If false, it is only counted how often a word occured `after` another word within the given window size}

\item{count.once}{Logical. A word can occur multiple times within the same window. If count.once is true, this is only counted as a single occurence (suggested).}
}
\value{
A graph in the Igraph format in which edges represent the adjacency of terms
}
\description{
A sliding window approach to calculate the co-occurence of words
}

